Refaturamento e aprimoração

+++ src/com/matheus/banco/model/pagamento/Boleto.java
@@ -2,20 +2,22 @@ package com.matheus.banco.model.pagamento;
 
 import com.matheus.banco.model.Pessoa;
 
+import java.math.BigDecimal;
+
 public class Boleto implements DocumentoPagavel, DocumentoEstornavel {
 
     private Pessoa beneficiario;
-    private double valor;
+    private BigDecimal valor;
     private boolean pago;
 
-    public Boleto(Pessoa beneficiario, double valor) {
+    public Boleto(Pessoa beneficiario, BigDecimal valor) {
         this.beneficiario = beneficiario;
         this.valor = valor;
         this.pago = pago;
     }
 
     @Override
-    public double getValorTotal() {
+    public BigDecimal getValorTotal() {
         return valor;
     }
 
@@ -28,7 +30,6 @@ public class Boleto implements DocumentoPagavel, DocumentoEstornavel {
     public void estornarPagamento() {
         pago = false;
     }
-
 
     @Override
     public void quitarPagamento() {
